// types/wishlist/enums.ts
export enum WishlistMessage {
  SUCCESS_GET = 'Lấy danh sách yêu thích thành công',
  SUCCESS_ADD = 'Thêm khóa học vào danh sách yêu thích thành công',
  SUCCESS_REMOVE = 'Xóa khóa học khỏi danh sách yêu thích thành công',
  SUCCESS_MOVE_TO_CART = 'Chuyển khóa học vào giỏ hàng thành công',
  FAIL_GET = 'Lấy danh sách yêu thích thất bại',
  FAIL_ADD = 'Thêm khóa học vào danh sách yêu thích thất bại',
  FAIL_REMOVE = 'Xóa khóa học khỏi danh sách yêu thích thất bại',
  FAIL_MOVE_TO_CART = 'Chuyển khóa học vào giỏ hàng thất bại',
  COURSE_ALREADY_IN_WISHLIST = 'Khóa học đã có trong danh sách yêu thích',
  COURSE_NOT_IN_WISHLIST = 'Khóa học không tồn tại trong danh sách yêu thích'
}